[FIX] MSVC complains about possible loss of information between size_t to int conversion.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@1614 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Wrapping/Matlab/btkMXAnalysis.cxx
@@ -73,7 +73,7 @@ void btkMXCreateAnalysisStructure(btk::Acquisition::Pointer acq, int nlhs, mxArr
           for (size_t i = 0 ; i < num ; ++i)
           {
             std::string& str = entryValues[inc][i];
-            str = (*it)->GetInfo()->ToString(i);
+            str = (*it)->GetInfo()->ToString((int)i);
             str = str.erase(str.find_last_not_of(' ') + 1);
             str = str.erase(0, str.find_first_not_of(' '));
           }
